FRD {FRD}	R Documentation

Bias-Aware Anderson-Rubin Confidence Sets

Description

This function calculates confidence sets for the parameter of interest in fuzzy regression discontinuity designs.

Usage


FRD(
  y,
  d,
  x,
  By = 0,
  Bt = 0,
  alpha = 0.05,
  cutoff = 0,
  search.int,
  subints = 3,
  diagnostic.plot = FALSE,
  eta = 1,
  bw.equal = TRUE,
  h = NA
)

Arguments

y
Outcome

d
Treatment status

x
Running variable

By
Bound on the second derivative of conditional expectation of the outcome function. It has to be positive. Default is zero.

Bt
Bound on the second derivative of the conditional probability of the treatment function. It has to be positive. Default is zero.

alpha
1-$\alpha$ is the confidence level of the confidence set.

cutoff
Cutoff value of the running variable. Default is zero.

search.int
The algorithm searches for the confidence set for the parameter of interest within this interval.

subints
Number of subintervals to search for the roots. It has to be an integer and greater than two.

diagnostic.plot
If TRUE, diagnostic plot of criterion function is drawn. Default is FALSE.

eta
eta specifies a maximal bound on w^max to improve finite sample coverage. Default is 1.

bw.equal
If FALSE, different bandwidths are used on each side of the cutoff. Default is TRUE.

h
If specific value is chosen, Anderson Rubin confidence sets are calculated with fix bandwidth. Default is NA.

tol
tol specifies which tolorance level to use to calculate confidence sets. Default is 10^-4

Value

A vector is returned which entries are the boundaries of the confidence set.

Examples


FRD(y=rnorm(100), d=rbinom(n=100, size=1, prob=0.5), x=rnorm(100), search.int=c(-1000,1000))

[Package FRD version 1.0.1 Index]