FRD-package {FRD}	R Documentation

Confidence Sets for fuzzy regression discontinuity designs

Description

This package provides a function to calculate bias-aware Anderson Rubin-type confidence intervals for the parameter of interest in fuzzy regression discontinuity designs. You can find a description of this function in Noack and Rothe (2020).

Details

The package includes the fuction FRD to calculate the confidence sets.

Author(s)

Claudia Noack <cnoack@mail.uni-mannheim.de> Christoph Rothe <crothe@mail.uni-mannheim.de>

References

Noack, C., & Rothe, C. (2020). Bias-aware inference in fuzzy regression discontinuity designs.

Examples


  ## Not run: 
    FRD(y=rnorm(100), d=rbinom(n=100, size=1, prob=0.5), x=rnorm(100), search.int=c(-1000,1000)) 
  
## End(Not run)

[Package FRD version 1.0.1 Index]