FRDsimple {FRD}	R Documentation

Bias-Aware Anderson Rubin Confidence Sets

Description

This function calculates simplified confidence sets for the parameter of interest in fuzzy regression discontinuity designs by assuming that the Confidence Set takes either the form (i) [a,b], (ii) \mathbb R or (-\infty, a] [b, \infty), where a<b.

Usage


FRDsimple(
  y,
  d,
  x,
  By = 0,
  Bt = 0,
  alpha = 0.05,
  cutoff = 0,
  tol = 1e-04,
  eta = 1,
  h = NA,
  bw.equal = TRUE
)

Arguments

y
Outcome

d
treatment status

x
running variable

By
Bound on the second derivative of conditional expectation of the outcome function. It has to be positive. Default is zero.

Bt
Bound on the second derivative of the conditional probability of the treatment function. It has to be positive. Default is zero.

alpha
1-$\alpha$ is the confidence level.

cutoff
Cutoff of the running variable

tol
tol Specifies which tolorance level to use to calculate confidence sets. Default is 10^-4

eta
eta specifies a maximal bound on w^max to improve finite sample coverage. Default is 1.

h
If specific value is chosen, Anderson Rubin confidence sets are calculated with fix bandwidth. Default is NA.

bw.equal
If FALSE, different bandwidths are used on each side of the cutoff. Default is TRUE.

Value

A vector is returned which entries are the boundaries of the confidence set.

[Package FRD version 1.0.1 Index]